% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimdo.R
\name{cimdo}
\alias{cimdo}
\title{CIMDO Dashboard}
\usage{
cimdo(obj = NULL, path = NULL, plotHeight = 800,
  viewerType = "paneViewer")
}
\arguments{
\item{obj}{object of cimdo class, Default: NULL}

\item{path}{character, path to CIMDO output directory, Default: NULL}

\item{plotHeight}{numeric, that sets the height of the plot output, Default 800}

\item{viewerType}{character, viewer to be used to launch the app to
c('paneViewer','dialogViewer','browserViewer'), Default: 'paneViewer'}
}
\description{
Launch shiny app UI for CIMDO from R console
}
\details{
pass to the function either a precompiled cimdo class object using
\code{\link{cimdo_read}} or pass a path to a CIMDO output directory to 
simultaneously read and launch.
}
\examples{
\dontrun{
if(interactive()){

PATH <- PATH_TO_CIMDO_OUTPUT

obj <- cimdo_read(PATH)

cimdo(obj)

}
}
}
