% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimdo_read.R
\name{cimdo_read}
\alias{cimdo_read}
\title{Create CIMDO object}
\usage{
cimdo_read(path)
}
\arguments{
\item{path}{character, path to CIMDO output directory}
}
\value{
data.frame
}
\description{
Read in CIMDO output directory files into a single data.frame
}
