% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_geom}
\alias{remove_geom}
\title{FUNCTION_TITLE}
\usage{
remove_geom(p, geom)
}
\arguments{
\item{p}{PARAM_DESCRIPTION}

\item{geom}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
