% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yq}
\alias{yq}
\title{FUNCTION_TITLE}
\usage{
yq(x, prefix = "\%Y", combine = "Q")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{prefix}{PARAM_DESCRIPTION, Default: 'Y'}

\item{combine}{PARAM_DESCRIPTION, Default: 'Q'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
